sunxi: display: Fix h/vsync TCON polarity bits to indicate active-high
Contrary to what is implemented in the Linux kernel, the U-Boot driver
supporting the sunxi TCON was assuming that the TCON0_IO_POL bits
indicate active-low signals when set for the hsync and vsync signals.

After looking at it with a logic analyzer, it appears that the signals
are active-low with the bits unset and active-high with the bits set.

Fix this by looking for the right flag (high instead of low) and
inverting the sync bits in the VIDEO_LCD_MODE declarations.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>