fb: add VMODE_DE_LOW flag to fb modeline
Some panels, like the LeMaker BL035 (3.5" for banana pi) need to have
their Display Enable signal pulled low. While this is already part of
the display_flags in flattened device tree, this change means that it
can also be used from the fbdev style modeline in the vmode field.

While this adds another flag to the FB_VMODE defines, there was a
precendent for this with FB_VMODE_ODD_FLD_FIRST for the imx.

Signed-off-by: Luc Verhaegen <libv@skynet.be>